requirements.txt
asttokens==2.4.1
bcrypt==4.2.0
cffi==1.17.1
cloudpickle==3.0.0
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fonttools==4.53.1
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
paramiko==3.5.0
parso==0.8.4
PIconnect==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
pysftp==0.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.2
pywin32==306
pyzmq==26.2.0
six==1.16.0
spyder-kernels==2.5.2
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
wcwidth==0.2.13
wrapt==1.16.0
